nineoneonemaps/requirements.txt
Django~=2.1.5
pinax-notifications~=5.0.3
djangorestframework~=3.9.1
django-import-export~=1.2.0
django-leaflet~=0.24.0
django-geojson~=2.12.0
django-social-share~=1.3.2
django-autoslug~=1.9.4
psycopg2-binary~=2.8.2
django_chartit~=0.2.9